tests/wpt/tests/css/css-borders/tentative/parsing/border-clip-computed.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Borders and Box Decorations 4 Test: Computed values of 'border-clip'</title>
<link rel="author" title="Sebastian Zartner" href="mailto:sebastianzartner@gmail.com">
<link rel="help" href="https://drafts.csswg.org/css-borders-4/#border-clip">
<meta name="assert" content="This test checks that the computed value of 'border-clip' is correct.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>

<div id="target"></div>

<script>
test_computed_value("border-clip", "normal");
test_computed_value("border-clip", "10px");
test_computed_value("border-clip", "10%");
test_computed_value("border-clip", "0");
test_computed_value("border-clip", "1fr");
test_computed_value("border-clip", "10px 10%");
test_computed_value("border-clip", "10px 1em 10% 1fr 2fr", "10px 16px 10% 1fr 2fr");
</script>
